worked up to course details

+++ lib/common/entities/widgets/base_text_widget.dart
@@ -4,7 +4,7 @@ import 'package:we_learn/common/entities/value/colors.dart';
 
 Widget reuseableText(String text,
     {Color color = AppColors.primaryText,
-    int fontSize = 16,
+    double fontSize = 16,
     FontWeight fontWeight = FontWeight.bold}) {
   return Container(
     child: Text(
